+++ react/UploadBlock.tsx
@@ -446,6 +446,7 @@ const UploadBlock: FunctionComponent<
               hiddenColumns={hiddenColumns ?? []}
               onReviewItems={onReviewItems}
               onRefidLoading={onRefidLoading}
+              backList={backList}
             />
             <div
               className={`mb4 mt4 flex justify-between ${handles.buttonsBlock}`}